This closes ticket #118.
Adds a new setting: MEDIA_SECURE_URL
This setting is put into the context as the media_url by the satchmo.shop.context_processor instead of the standard MEDIA_URL when the page is being served securely.  In addition, a new context variable "is_secure" is also populated.

This patch updates the google tags to support secure-aware templates, and updates the templates to use the newly security-aware versions of the tags.

Lastly, because it was tangled up with my other Google updates, this patch adds support for google adwords and google signup conversion tracking.  It adds these tags to the registration template.



git-svn-id: svn://satchmoproject.com/satchmo/trunk@636 a38d40e9-c014-0410-b785-c606c0c8e7de